+++ lib/src/mcu/scheme/scheme_rainbow.dart
@@ -29,6 +29,11 @@ class SchemeRainbow extends DynamicScheme {
     required Hct sourceColorHct,
     required super.isDark,
     required super.contrastLevel,
+    Hct? secondarySourceColorHct,
+    Hct? tertiarySourceColorHct,
+    Hct? neutralSourceColorHct,
+    Hct? neutralVariantSourceColorHct,
+    Hct? errorSourceColorHct,
   }) : super(
           sourceColorArgb: sourceColorHct.toInt(),
           variant: Variant.rainbow,
@@ -36,21 +41,35 @@ class SchemeRainbow extends DynamicScheme {
             sourceColorHct.hue,
             48.0,
           ),
-          secondaryPalette: TonalPalette.of(
-            sourceColorHct.hue,
-            16.0,
-          ),
-          tertiaryPalette: TonalPalette.of(
-            MathUtils.sanitizeDegreesDouble(sourceColorHct.hue + 60.0),
-            24.0,
-          ),
+          secondaryPalette: secondarySourceColorHct != null
+              ? TonalPalette.of(
+                  secondarySourceColorHct.hue,
+                  16.0,
+                )
+              : TonalPalette.of(
+                  sourceColorHct.hue,
+                  16.0,
+                ),
+          tertiaryPalette: tertiarySourceColorHct != null
+              ? TonalPalette.of(
+                  tertiarySourceColorHct.hue,
+                  24.0,
+                )
+              : TonalPalette.of(
+                  MathUtils.sanitizeDegreesDouble(sourceColorHct.hue + 60.0),
+                  24.0,
+                ),
           neutralPalette: TonalPalette.of(
-            sourceColorHct.hue,
+            neutralSourceColorHct?.hue ?? sourceColorHct.hue,
             0.0,
           ),
           neutralVariantPalette: TonalPalette.of(
-            sourceColorHct.hue,
+            neutralVariantSourceColorHct?.hue ?? sourceColorHct.hue,
             0.0,
           ),
+          customErrorPalette: errorSourceColorHct == null
+              ? null
+              : TonalPalette.of(
+                  errorSourceColorHct.hue, errorSourceColorHct.chroma),
         );
 }